requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.5.0
streamlit>=1.28.0
qrcode[pil]>=7.4.2

# Optional: For fine-tuning demo (REQUIRES NVIDIA/AMD/Intel GPU - NOT macOS compatible)
# 
# Installation steps (on Linux/Windows with GPU):
# Step 1: Install base packages
#   pip install torch transformers trl datasets bitsandbytes
#
# Step 2: Install unsloth (skip xformers on macOS)
#   pip install "unsloth[colab-new]" --no-deps
#   pip install unsloth_zoo
#   pip install diffusers torchvision
#
# IMPORTANT NOTES:
# - Unsloth ONLY works on NVIDIA, AMD, or Intel GPUs (NOT Apple Silicon)
# - macOS users: Fine-tuning demo will not work due to GPU requirements
# - bitsandbytes version conflicts may occur on macOS (this is expected)
# - The RAG demo works perfectly on all platforms without GPU
#
# For macOS: Only install core dependencies above (RAG will work fine)

